feat: add more `window api` example
